rename to `contract_upgrade.rs`
