+++ crates/contract/tests/sandbox/mod.rs
@@ -1,5 +1,5 @@
-pub mod back_compatibility;
 pub mod ckd;
+pub mod contract_upgrade;
 pub mod integration_tee_cleanup_after_resharing;
 pub mod sign;
 pub mod tee;
